Refactor: 리뷰 반영사항 추가
