+++ app/users/[id]/_components/editable-image.tsx
